Creating update_charts.py
